requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
arrow==1.4.0
Authlib==1.6.8
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
codecarbon==3.2.3
colorama==0.4.6
cryptography==46.0.5
diffprivlib==0.6.6
idna==3.11
joblib==1.5.3
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
nvidia-ml-py==13.590.48
pandas==3.0.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
questionary==2.1.1
RapidFuzz==3.14.3
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
